feat: move convert function to helpers
