+++ src/modules/skills/index.ts
@@ -1,0 +1,1 @@
+export { default as SkillsList } from "./ui/skills-list";
